server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
asyncpg==0.30.0
apscheduler==3.10.4
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.4
